Backend/requirements.txt
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.12.2